# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing Personnel On Board, Vessels, and Rooms.",
        "title": "POB Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/vessels": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches a list of all vessels and installations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vessels"
                ],
                "summary": "List all vessels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a physical vessel or installation to track",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vessels"
                ],
                "summary": "Create a new vessel",
                "parameters": [
                    {
                        "description": "Vessel Details",
                        "name": "vessel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateVesselInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/vessels/{id}/pob": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fast redis-based query for real-time Personnel On Board",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vessels"
                ],
                "summary": "Get live POB for a vessel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vessel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/vessels/{vesselId}/rooms": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets up a physical room with beds",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Create a room inside a vessel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vessel ID",
                        "name": "vesselId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Room Details",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.CreateRoomInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/vessels/{vesselId}/rooms/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assigns someone to a room and inherently increments the real-time POB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Assign personnel to a room bed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vessel ID",
                        "name": "vesselId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Assignment logic",
                        "name": "room",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.AssignRoomInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.AssignRoomInput": {
            "type": "object",
            "required": [
                "personnel_id",
                "room_id",
                "starts_at",
                "type",
                "vessel_id"
            ],
            "properties": {
                "activity_id": {
                    "description": "Required if Type == \"Flexible\"",
                    "type": "string"
                },
                "ends_at": {
                    "description": "Required if Type == \"Flexible\"",
                    "type": "string"
                },
                "offshore_role_id": {
                    "description": "Required if Type == \"Core\"",
                    "type": "string"
                },
                "personnel_id": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "type": {
                    "description": "\"Core\" or \"Flexible\"",
                    "type": "string"
                },
                "vessel_id": {
                    "type": "string"
                }
            }
        },
        "service.CreateRoomInput": {
            "type": "object",
            "required": [
                "code",
                "name",
                "vessel_id"
            ],
            "properties": {
                "capacity": {
                    "type": "integer",
                    "minimum": 1
                },
                "category": {
                    "description": "\"Dedicated\" vs \"Transient\"",
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "deck": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "vessel_id": {
                    "type": "string"
                }
            }
        },
        "service.CreateVesselInput": {
            "type": "object",
            "required": [
                "code",
                "name",
                "type"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "minimum_safe_pob_capacity": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pob_capacity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}